build(ci): add bitrise step `cocoapods-install@2`
build(ci): add bitrise steps for firebase cli installation

build: fix `assemble.sh` fails when .build dir is absent

build: fix pod setup

build(ci): update bitrise machine stack

build: remove Pods_SmartInvesting.framework from PBXBuildFile & PBXFrameworksBuildPhase

build(ci): re-arrange cocoapods-install@2

+++ BuildScripts/assemble.sh
@@ -1,7 +1,11 @@
 #!/bin/sh
 set -e
 
-find .build -mindepth 1 -delete
+if [ -d ".build" ]; then
+    find .build -mindepth 1 -delete
+else
+    echo ".build directory does not exist"
+fi
 
 xcodebuild \
   -workspace SmartInvesting.xcworkspace \
